+++ packages/vda-common-test/src/index.ts
@@ -1,2 +1,4 @@
 export * from './const'
 export * from './utils'
+export * from './erc20/erc20Manager'
+export * from './storage_node'