+++ packages/vda-common-test/CHANGELOG.md
@@ -1,4 +1,8 @@
+2023-11-24 (v2.3.1)
+-------------------
+- Update type of `nonce` from `number` to `BigNumberish`
+- Add `ERC20` contract manager api
+
 2023-04-20 (v2.3.0)
 -------------------
-
 - Initial implementation